feat: allow list of files instead of Doxygen theme
Allows users to specify a list of custom theme files which will be added
to `DOXYGEN_HTML_EXTRA_STYLESHEET`.

If any of the files are not found, the theme is reverted to "original".
It reverts to "original" and not "awesome", because (presumably) the
user still wishes to use the custom theme and just mistyped the path, so
there is no need to download the `awesome` theme.

Note: In the `project_options` macro the `DOXYGEN_THEME` flag was made
to take multiple parameters for this to work.