Change to find dependencies in Module mode
If dependencies are not found in Module mode, the search falls back to
Config mode.

However, when explicitly use find_package(<name> Config) as we used to do.
The search won't fall back to Module mode.

If Config mode is preferred, use CMAKE_FIND_PACKAGE_PREFER_CONFIG to
change the behaviour.