fix(MSVC): Don't enforce /Zi if POLICY CMP0141 is available
With the current implementation if CMAKE_MSVC_DEBUG_INFORMATION_FORMAT is set,
it won't be respected as the /Zi flag is always set in enable_sanitizers.

If POLICY CMP0141 is available, don't set the /Zi flag,
instead use MSVC_DEBUG_INFORMATION_FORMAT, if the user didn't set
CMAKE_MSVC_DEBUG_INFORMATION_FORMAT or set it to EditAndContinue,
than only in that case set it to ProgramDatabase.